pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "photo-sync"
dynamic = ["version"]
description = "Sync photos from Apple Photos (and more) to static site galleries"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "N2 Solutions", email = "nick@n2solutions.io" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "click>=8.1",
    "osxphotos>=0.68",
    "Pillow>=10.0",
    "rich>=13.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "pre-commit>=4.0",
]

[project.scripts]
photo-sync = "photo_sync.cli.main:cli"

[project.urls]
Homepage = "https://github.com/n2solutionsio/photo-sync"
Repository = "https://github.com/n2solutionsio/photo-sync"
Issues = "https://github.com/n2solutionsio/photo-sync/issues"

[tool.hatch.version]
path = "src/photo_sync/__about__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/photo_sync"]

[tool.ruff]
target-version = "py310"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "osxphotos.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tomli"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[tool.coverage.run]
source = ["photo_sync"]

[tool.coverage.report]
show_missing = true
